% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{dataLinks}
\alias{dataLinks}
\alias{dataLinks.default}
\alias{dataLinks.character}
\alias{dataLinks.list}
\title{web pages}
\usage{
dataLinks(x = NULL, ...)

\method{dataLinks}{default}(x = NULL)

\method{dataLinks}{character}(x = NULL)

\method{dataLinks}{list}(x = NULL, verbose = NULL)
}
\arguments{
\item{x}{character vector of Content (means HTML text)}

\item{x}{list containing Content (means HTML text)}
}
\value{
list of URL links

list of URL links
}
\description{
web pages
}
