% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{htmlContent}
\alias{htmlContent}
\alias{htmlContent.default}
\alias{htmlContent.xml_node}
\title{html content, reads one}
\usage{
htmlContent(x, ...)

\method{htmlContent}{default}(x)

\method{htmlContent}{xml_node}(x = NULL, verbose = NULL)
}
\arguments{
\item{x}{xml_document or xml_node}
}
\value{
one large characer vector element
}
\description{
html content, reads one
}
