% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{readHTML}
\alias{readHTML}
\alias{readHTML.default}
\alias{readHTML.character}
\alias{readHTML.list}
\title{reads body of resulting URL and returns a ("xml_document" "xml_node")}
\usage{
readHTML(x = NULL, ...)

\method{readHTML}{default}(x = NULL)

\method{readHTML}{character}(x = NULL, verbose = NULL)

\method{readHTML}{list}(x = NULL, verbose = NULL)
}
\arguments{
\item{x}{character vector of strings that are urls}

\item{x}{list of single element character vectors of url strings}
}
\value{
list of ("xml_document" "xml_node")

list of ("xml_document" "xml_node")
}
\description{
reads body of resulting URL and returns a ("xml_document" "xml_node")
}
