% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{returnsBankrutcyFilings}
\alias{returnsBankrutcyFilings}
\title{runs the main}
\usage{
returnsBankrutcyFilings()
}
\value{
xts bankruptcy filings
}
\description{
runs the main
}
\examples{
\dontrun{
EXAMPLE
# EXAMPLE
}
}
