% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UScourtsBankrutcyFilings.R
\docType{package}
\name{UScourtsBankrutcyFilings}
\alias{UScourtsBankrutcyFilings}
\alias{UScourtsBankrutcyFilings-package}
\title{UScourtsBankrutcyFilings: Retrieve or update US Federal Court Bankruptcy Filings}
\description{
Collect or update US Federal Court Bankruptcy Filings.
}
