% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilary.R
\name{webPages}
\alias{webPages}
\alias{webPages.default}
\alias{webPages.character}
\alias{webPages.list}
\title{web pages}
\usage{
webPages(x = NULL, ...)

\method{webPages}{default}(x = NULL)

\method{webPages}{character}(x = NULL)

\method{webPages}{list}(x = NULL, verbose = NULL)
}
\arguments{
\item{x}{url of a character vector}

\item{x}{list containing (xml_document or xml_node)}
}
\value{
list of large "character" elements

list containing large "character" elements
}
\description{
web pages
}
